Fix error message formatting
git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@4415 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf

+++ trunk/libmesh/src/systems/fem_system.C
@@ -288,7 +288,7 @@ void FEMSystem::assembly (bool get_residual, bool get_jacobian)
                                 << " detected in analytic jacobian on element "
                                 << _femcontext.elem->id()
 			        << ", side "
-                                << _femcontext.side << '!' << std::endl;
+                                << static_cast<unsigned int>(_femcontext.side) << '!' << std::endl;
 
                   unsigned int old_precision = libMesh::out.precision();
                   libMesh::out.precision(16);
